Some little issues
**I open again the same pull request on the right repository**
PLEASE TRY OUT THE CODE BEFORE MERGE

FWD:

Good evening Pavel, working on the assignement I've found some "issues" on the lab7

If I well understood, a PUT request is supposed to UPDATE a resource, or if that resource does not exists, the system should CREATE a new one. Right?

Supposing that the DB is not containing the ID x (x = some ID value not contained in the person table) following the code flaw for the following request:

PUT http://127.0.0.1:5700/sdelab/person/123 (with a proper body) 

The  function putPerson(Person person) (line 69) will be called..
then a new person wil be created with an id provided by the DB (line 72) 
then getPersonById(this.id) will be called (line 76)  the ID (as supposed before) won't be in the DB.. So the variable person at line 106 wil be null 
then the code will breack at line 107 because the person.toString() method is called on a NULL variable. 

Now suppose the ID is present into the DB, then unfortunately a person object will still be created  (with an ID given by the DB check line 72) and then 
the existing entry with the given ID will be updated. (this is not correct). 

I edited the code in order to make it work properly, please have a look to the comments. 
Please try  to run it before merge the pull request, although I've tested the code, I've also added some println in-line here in GitHub... 

Maybe I've not been clear anyway for any clarification my nick on slack is andrea_g
I do apologize for my "school" English..

Thanks a lot for the hard work you do for us! 

Sincerely 
Andrea Galloni